+++ lib/features/games/basketball/nba/widgets/matchup_card/matchup_card.dart
@@ -180,7 +180,7 @@ class MatchupCard extends StatelessWidget {
                                                   gameName: state.league,
                                                 ),
                                                 spread: gameData.overUnder
-                                                    .toDouble(),
+                                                    .toDouble() as double,
                                                 mainOdds: gameData.overPayout
                                                     .toString(),
                                                 betType: Bet.tot,
@@ -314,7 +314,7 @@ class MatchupCard extends StatelessWidget {
                                                     .toString(),
                                                 betType: Bet.tot,
                                                 spread: gameData.overUnder
-                                                    .toDouble(),
+                                                    .toDouble() as double,
                                                 awayTeamData:
                                                     state.awayTeamData,
                                                 homeTeamData:
